Add server channel, member, and section components
The code introduces new components for server channel, member and section. It updates the server sidebar to present the channels and members as part of the server view. A sepatator component was inserted for visual layout arrangement. Also, adjustments are made on modal-store hooks and create-channel-modal to support different channel types.